Implement scrolling

+++ kernel/include/console.hpp
@@ -14,11 +14,11 @@
 
 void init_console();
 
-void set_column(long column);
-long get_column();
+void set_column(size_t column);
+size_t get_column();
 
-void set_line(long line);
-long get_line();
+void set_line(size_t line);
+size_t get_line();
 
 void wipeout();
 void k_print(char key);
